LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/resource-timing.https-expected.txt

Blocked access to external URL https://www1.localhost:9443/service-workers/service-worker/resources/missing.jpg
Blocked access to external URL https://www1.localhost:9443/service-workers/service-worker/resources/missing.jpg


FAIL Controlled resource loads assert_unreached: unexpected rejection: assert_greater_than: Generated response expected a number greater than 0 but got 0 Reached unreachable code
PASS Non-controlled resource loads 
